Reestructuración de routing
- Se cambió la estructura del routing en app.jsx
- Se cambió la estructura de /home, para poder renderizar el navbar en todas las rutas hijas
- Se crea HomeNavigation.jsx como navbar
- Se refactoriza Home.jsx para renderizar el contenedor de servicios
