+++ Multiserv/client/src/pages/CreateService/CreateService.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const CreateService = () => {
+    return (
+        <div>
+            crear servicio
+        </div>
+    )
+}
+
+export default CreateService
