+++ Multiserv/client/src/pages/Home/Home.jsx
@@ -1,25 +1,14 @@
-/* Atoms */
-import LinkTo from '../../Components/Atoms/LinkTo/LinkTo'
-/* Organisms */
-import Nav from "../../Components/Organisms/NavBar/Nav"
-/* Assets */
-import Img from '../../assets/Icons/profile.png'
-/* Icon */
-import { AiFillHome, AiFillCalendar, AiFillStar, AiOutlineLoading3Quarters } from "react-icons/ai";
-import { BsFillChatDotsFill, BsSortAlphaDownAlt, BsSortAlphaUpAlt, BsSortDown, BsSortDownAlt } from "react-icons/bs";
-import { MdEdit, MdNotifications, MdFavorite, MdHomeRepairService } from "react-icons/md";
-import { FaUserAlt } from "react-icons/fa";
-import { BiLogOut } from "react-icons/bi";
-/* React Hooks */
+
+import { AiFillStar, AiOutlineLoading3Quarters } from "react-icons/ai";
+import { BsSortAlphaDownAlt, BsSortAlphaUpAlt, BsSortDown, BsSortDownAlt } from "react-icons/bs";
+
 import { useEffect, useState } from 'react';
 /* React redux */
 import { useSelector, useDispatch } from 'react-redux'
 import { buscar, orderAlph, orderRating, services, resetOrder, filterCats, getCats } from '../../redux/actions/actions';
 import ServiceCard from '../../Components/Molecules/ServiceCard/ServiceCard';
 import { useNavigate } from 'react-router';
-import s from "../../Components/Organisms/UserProfile/UserProfile.module.css"
 import Input from '../../Components/Atoms/Input/Input';
-import ButtonXartiago from '../../Components/Atoms/ButtonXartiago/ButtonXartiago';
 import ListBox from '../../Components/HeadLess/ListBox/ListBox'
 
 const Home = () => {
@@ -27,7 +16,6 @@ const Home = () => {
     const servicios = useSelector((state) => state.servicios)
     const categorias = useSelector((state) => state.categories)
     const navigate = useNavigate();
-    const [verPerfil, setVerPerfil] = useState(false)
     const [buscador, setBuscador] = useState('')
     const [order, setOrder] = useState(null)
     const [filter, setFilter] = useState(null)
@@ -82,23 +70,8 @@ const Home = () => {
         setFilter(obj)
     }
 
-
-    var foto = Img
-    if (localStorage.length > 0 && datosSesionFromLocalStorage.photoURL) {
-        foto = datosSesionFromLocalStorage.photoURL
-    }
     // si necesitan datos de la sesión se encuentran en la variable datosSesionFromLocalStorage
-    const arr = [
-        <LinkTo linkClass='m-4 flex justify-center' page='home' render={<AiFillHome size='28' color='white' />} />,
-        <LinkTo linkClass='m-4 flex justify-center' page='home/chat' render={<BsFillChatDotsFill size='28' color='white' />} />,
-        <LinkTo linkClass='m-4 flex justify-center' page='home/profile' render={<FaUserAlt size='28' color='white' />} />,
-        <LinkTo linkClass='m-4 flex justify-center' page='home/schedule' render={<AiFillCalendar size='30' color='white' />} />
-    ]
-    const logout = (e) => {
-        e.preventDefault();
-        localStorage.removeItem("datoSesion")
-        navigate("/")
-    }
+
     const handleListValue = (obj) => {
         setOrder(obj)
     }
@@ -141,76 +114,9 @@ const Home = () => {
         },
         ...categorias
     ]
-    const handleClick = () => {
-        setVerPerfil(!verPerfil);
-    }
-
-    const validarLogitudNombre = (nombre) => {
-        let nombres = nombre.split(" ")
-        return nombres;
-    }
-
-    var name = " Invitado"
-    var email = ""
-    if (localStorage.length > 0 && datosSesionFromLocalStorage.displayName) {
-        name = datosSesionFromLocalStorage.displayName
-        email = datosSesionFromLocalStorage.email
-    }
-
-    const resultadoNombre = validarLogitudNombre(name)
-    const modal = (
-        verPerfil ?
-            <div className={s.UserProfile__OnClick}>
-                <div className="flex w-full border-b-2 py-5 pr-2">
-                    <img className="mx-2" src={foto} alt="" />
-                    <div className="flex flex-col justify-center truncate">
-                        <span className="font-semibold text-xl">{`${resultadoNombre[0]} ${resultadoNombre[1]}`}</span>
-                        <span className="text-sm text-gray-500 -mt-1">{email}</span>
-                    </div>
-                </div>
-                {
-                    email !== "" &&
-                    <div className="flex flex-col w-full justify-center py-2">
-                        <div className="flex my-1 items-center pl-3">
-                            <MdEdit className="mr-2" />
-                            <span className="font-semibold">Editar Perfil</span>
-                        </div>
-                        <div className="flex my-1 items-center pl-3">
-                            <MdNotifications className="mr-2" />
-                            <span className="font-semibold">Notificaciones</span>
-                        </div>
-                        <div className="flex my-1 items-center pl-3">
-                            <MdFavorite className="mr-2" />
-                            <span className="font-semibold">Lista Favoritos</span>
-                        </div>
-                        <div className="flex my-1 items-center pl-3">
-                            <MdHomeRepairService className="mr-2" />
-                            <span className="font-semibold">Ser Provedor</span>
-                        </div>
-                    </div>
-                }
-                {datosSesionFromLocalStorage ? (<div className="flex items-center justify-center mt-2 w-2/5 mt-3"><button onClick={logout} className="font-semibold text-gray-50 flex w-full flex-nowrap bg-green-700 p-2 py-2 px-4 justify-center items-center rounded-md">Log out</button></div>) : (<ButtonXartiago
-                    btn="Regresar"
-                    page=""
-                    clase="w-2/5 mt-3"
-                    btnClass="font-semibold text-gray-50 flex w-full flex-nowrap bg-green-700 p-2 py-2 px-4 justify-center items-center rounded-md"
-                />)}
-            </div>
-            :
-            null)
-
 
     return (
-        <div className="flex">
-            <Nav
-                clase='w-20 h-screen p-4 pt-6 flex flex-col justify-between justify-center bg-blue-900'
-                imgClass='w-16 rounded-full cursor-pointer'
-                imgOnClick={handleClick}
-                imagen={foto}
-                imgName='Logo'
-                arr={arr}
-            />
-            {modal}
+        <>
             {
                 loading ? (
                     <div className="w-full flex flex-col h-screen justify-center items-center">
@@ -258,7 +164,7 @@ const Home = () => {
                     </div>
                 )
             }
-        </div>
+        </>
     )
 }
